Handle pagination/load events.
Included AJAX call that would hit a different route on the Rails API that passed in the number of images currently on the screen as the results offset, in addition to the search object ID in order to query the database for that particular search object's results. Potential bug identified: because of the bug mentioned before regarding an undefined method error, the load more functionality may not work given that the GET requests does not always seem to work. Will troubleshoot given time.